Added obstacles and AI navigation
I made static and dynamic obstacles for the player and enemy to navigate around. I also applied AI navigation to the enemy movement script so that I chases the player object. I added a navigation map on the ground of the places that the enemy can go. Added text for when the player hits the enemy and loses. Added the lose condition.