Audio
I added audio to my game! I added background music to a blank audio source. I also added a "collect" sound when you collect a pick-up that comes from the player. Another audio was the  game over audio when the player loses, it is stored in the enemy object but called in the player script. Lastly, I added a win sound to play from the player when all the pick-ups are collected.