Visual FX
I added a visual effect for when the player collects a pick up item and it uses burst particles. They come from where the pickup item was. I also added an explosion that appears when the enemy catches the player. I added a texture sheet to make it more realistic smoke shapes and added a color gradient. I could not figure out how to make the effects work unless I had Play on Awake checked, which is not ideal but that's what the tutorial said to do.